string tokenizer
The string tokenizer was horribly broken and now it is slightly less
broken.  Basically the string tokenizer function, used for breaking up
arrays returned by the transaction, is not a very intelligent tokenizer.
Currently the string tokenizer are making assumptions based on the
format of the column data that are in arrays.  They may also be making
assumptions based on PostgreSQL specific formatting.

In fixing up the tokenizer, also fixed up the usage of the tokenizer and
the debug output.