pgsql transaction logic updated to v1.9.0
Major changes to updated the PostgreSQL related code to v1.9.0.

Debugging code was been completed and additional code added to help
validate the transaction logic.  Updated the formatting look of the
outpout to more easily read what is going on.

Additional checks are made to verify the number of the items returned in
column array match what the database transactions says there are.
Verbosity of individual checks still vary.  There are certain elements
that the spec does not guarantee the number of elements for.  These
particular pieces are not checked and are marked FIXME in case this
interpretation is incorrect.

Fixed string tokenizer to correctly parse string.  Also added another
tokenizer to parse arrays out of arrays.  Should probably be combined at
some point.  There are still potential strings that will break the
tokenizer, but it is not expected that EGen will create those types of
strings.  Exceptions are noted in the code.  Move the tokenizer code
from the TxnBasedB.h header file to the transaction.h header file.  It
is supposed to be helper inlined functions as opposed to a class method.

Several bugs fixed the PostgreSQL stored functions.  All stored
functions should be updated to spec, except for the Market Feed, which
is noted in the code.  For the PostgreSQL stored functions (not the
pl/psql code):

* Fixed various data size problems when various id's were set to int.
  All have been fixed to be long's
* Fixed a incorrectly sized news variable in Security Detail frame 1.
* Properly return day data in Security Detail frame 1.
* Fixed a bug in Trade Lookup frame 1 when checking is_cash.
* Properly return trade history data in Trade Lookup Frame 1.
* Fixed the handling of trade history data in Trade Update frame 1.
* Trade update frame 3 was not escaping a string correctly for one of
  its queries.
* Fixed queries in the Trade Result frame to handle escaped strings, and
  looping through return values.
* Added error check for the Data Maintenance frame in the case where the
  table name doesn't match.

The driver code has been updated to match the updated stored functions.
The order of reading in the values from the stored functions are mostly
ordered by input/output order of the transaction descriptions.  The
intent is to be able to easily proof things against the specification.

* Use strncpy over strcpy to prevent possible memory violations cause by
  bad data returned by the database.
* Correctly parse history output in the Customer Position frame 2.
* More correctly implement the Market Feed frame.  Limitations in the
  ability to use transactions in PostgreSQL stored function limits the
  implementation for that database.
* Correctly parse the fin and dts data in the Security Detail frame 1.
* Correctly parse trade history data in all of the Trade Lookup frames.
* Fixed memory violation bug caused by misuse of sscanf in the Trade
  Update frames when reading timestamps.
* The correct isolation level was not set to REPEATABLE READ for all of
  the possible Trade Update frame transaction starting locations.
* Fixed MarketFeed to return proper status code on error.
* Fixed call to Security Detail frame 1 to use a DATE instead of a
  TIMESTAMP datatype.
* Removed COMMIT call in Trade Result frame 5 that should not be there.