remove more dynamic memory allocation
Removed new/delete calls relating to the TalkToSUT.  In the CESUT
class, the structure used to send data to and from the sut was being
dynamically allocated per call.  Now a static class variable is reused
without allocated memory for it.

In the case of the MEESUT class, a static variable is not declared
per call as opposed to dynamically allocated per call.

I don't know if this is a better memory management move or not, but
cross your fingers.