smarter TalkToSUT retuen code handling
Removed logging __FILE__ and __LINE__ macros when there is a socket
error because it is just telling us where the logging call was made, not
where the error was actually caught.

Modified the the TalkToSUT return a meaningful number.  Either a -1 to
let the caller know there was an error, or the length of the data
received from the sut.

Finally, the caller for TalkToSUT now pays attention to the return code
to TalkToSUT and returns True only if there were no problems (as far as
the code can tell) and False is there was a problem.  Previously the
code always returned True no matter what.