autoconf pg_config and pkg-config removal
Stop using pg_config and pkg-config to set the proper include flags and
linker flags.  The reason for this is we want to use the autoconf macros
in order for the linker flags to get appending into the correct
location, because location matters for things like --libpqxx.  What this
breaks is specifying non standard installation paths.  This will be
reincorporated correctly later.