database code encapsulation phase 2
Encapsulated all the handling of data returned by the database calls
into the DBConnection class object.  The intent is to keep all data
specific api calls in one location, which I hope is beneficial to
porting to work with other databases in the future.

The types of calls that are encapsulated are the database SQL execution
calls, setting of isolation levels, and start and end transaction
statements.