fix error handling
In the PostgreSQL related code, fixed to throw errors when encountered
while executing the transaction frames.  This is required to stop the
transaction from executing.  Otherwise the driver will continue trying
to execute frames in the transaction and unexpected behavior will occur.

In the event of errors thrown from the DBConnection class to abort the
transaction, the BrokerageHouse now traps and handles setting the
transaction status to rollback properly.  The trapping of errors thrown
here is what actually stops the transaction.