update postgresql scripts
Allow DBNAME to be set via the command line in addition to the
environment variable.  Will try to move all the environment variables in
this direction.

Corrected how the customer parameters are used with EGen.  Specifically
the starting position, total customer and number of customers in the
instance.  Also set the default parameters to the proper minimum values.

Don't run 'create_lang plpgsql' anymore.  It's been loaded automatically
prior to PostgreSQL 9.0 so the scripts actually need to be fixed to
check the version of PostgreSQL before loading them.  The truth is the
kit doesn't need it unless the pl/pgsql stored functions are used.

Reorganized the parameter flags to be more consistent with the kit.

Updated the script to collect the SQL plans to match the C stored
functions more closely.

Updated the database stats collection functions to collect data a little
more uniformly.

Updated README-POSTGERSQL to match.