use COPY for pgsql loader
This change also hopefully fixed an elusive bug.  The loader failed to
load a database with a 10,000 customer cardinality.  Don't know if it's
because of libpqxx, but noting it here anyway.

Using COPY via pipes to psql instead of INSERT and the libpqxx library.
This should be a huge performance increase because with transactions the
WAL log is bypassed with these more recent version of PostgreSQL.  This
simplifies the code a little bit.

An unfortunate side effect is that EGen is not passing arguments with
the -p flag properly when they contain dashes (-) in the argument.  This
does not allow us to pass command line arguments to psql.  As a fall
back, we now rely on the environment variable to be set in order for
EGen to work.  The scripts for PostgreSQL attempts to help manage that.