refactor driver thread failure handling and error messages
If there is a problem starting the data maintenance thread, write to
cerr instead of to the log file before exiting.

When a thread is exiting, write notice to cout instead of error log,
since it's not an error.

When the data maintenance thread is starting, write to cout instead of
the error log, since it's not an error.