fix direct loading and make it default again
EGenLoader appears to be partitioning the data generation, in a way, for the
growing tables.  We originally thought it was loading everything at once, such
that we could do the TRUNCATE and FREEZE trick with PostgreSQL.  But it turns
out that EGenLoader would load multiple parts, thus the next part would
TRUNCATE the previously loaded data.