pgsql pl/c remove superfluous input logging
The debug code dumps the input parameters to elog() at the top of each
function, but probably shouldn't need to dump it again every time the frame
fails.

I think this was initially done at a time when the initial dump of of the input
variables was not necessarily done but if a frame failed, it would guarantee
the input parameters were dumped.

So if the input parameters need to be captured, this should be done by a
combination of postgres logging options (to verify what postgres sees) and
building the C functions with -DDEBUG (to verify the C code is handling the
parameters correctly.)