Replace vector::iterator with time_t
Not sure if this is the root issue, but I have more confidence in checking if
`j < vAux.size()` over `p != vAux.end()`.

Also add the second condition to prevent looping past the size of the various
pOut arrays, but just for some of the time stamp arrays.

At least it's less lines of code, assuming everything else is equivalent.