pgsql: set default load type to FLAT
There is a bug in the CUSTOM data generation at higher customer counts,
somewhere larger than 5,000 and lower than 100,000 that is producing
inconsistent data regardless of whether the data is chunked.  A foreign
key will fail to create because of a missing key but table cardinality
looks sane otherwise.

Set the default load type to FLAT for the time being.